I hope this is the beginning of the end
